requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.5
aiosignal==1.4.0
astunparse==1.6.3
attrs==23.2.0
audioop-lts==0.2.1
braceexpand==0.1.7
certifi==2024.6.2
chardet==3.0.4
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
frozenlist==1.4.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
import_expression==1.1.5
jishaku==2.5.2
legacy-cgi==2.6.3
multidict==6.0.5
pillow==11.3.0
propcache==0.3.2
python-dotenv==1.0.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
wheel==0.43.0
yarl==1.9.4
